Memoria Set1 benchmark. It's only about 6 times slowler than PackedTree<BigInt, EmptyValue>.

+++ src/tools/benchmarks/pset_find.hpp
@@ -49,7 +49,7 @@ template <typename List> class ListExecutor;
 template <Int BranchingFactor>
 class PSetBenchmark: public BenchmarkTask {
 
-	typedef PMapFindTypes<Int, EmptyValue, 1, BranchingFactor> 	Types;
+	typedef PMapFindTypes<BigInt, EmptyValue, 1, BranchingFactor> 	Types;
 
 	typedef typename Types::Accumulator		Accumulator;
 	typedef typename Types::Key				Key;
